Separar UI de control de compras
